+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Adminer dev
+### Plugins
+- New method assetUrl() returning the URL of a static file in the compiled Adminer
 
 ## Adminer 6.0.1 (released 2026-08-14)
 - Run the development version from the adminer/ directory alone, copied anywhere and under any name